Solved the issues with libraries not loading from /usr/lib.
Included a ld.so.conf with directories /usr/lib and /lib. Then, at boot time in /etc/init.d/openvswitch, ldconfig is called so the cache is updated (only needed for the first time and after any library installation). Look for a better place to do it. As a side effect, we no longer need libssl (openssl libraries) to be installed at /lib (/usr/lib is just fine).