confine.disk-parted: Move partition preparation to a function.
This enables calling the script with a "post" argument for it to complete a
previous run which was left incomplete because of a required reboot to let the
kernel get the right partition size on machines which don't get the right disk
geometry.  It also avoids code duplication.  The preparation of the fstab has
also been moved after partition preparation so that it doesn't get in the way
if the preparation succeeded partially or the filesystems where available from
a previous partitioning.

This should resolve issue #185.